requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
torch==2.1.1
prophet==1.1.5
neuralprophet==0.6.0
python-dotenv==1.0.0
pydantic==2.5.2
joblib==1.3.2
plotly==5.18.0
pytest==7.4.3
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
lightgbm==4.1.0
xgboost==2.0.2
optuna==3.4.0 